requirements.txt
Django==3.0
django-crispy-forms==1.8.1
matplotlib==3.1.2
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
tensorflow==2.0.0

